# www/api-transfer/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger-Demo",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getAll",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "list of users",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "document": {
                                                        "type": "string"
                                                    },
                                                    "profile": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "description": "Return created user",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "document": {
                                        "type": "string"
                                    },
                                    "profile": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "data of new user",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "document": {
                                                    "type": "string"
                                                },
                                                "profile": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "The given data was invalid",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Erro",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/users/{user_id}/purse": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get purse by user",
                "description": "Returns purse by user",
                "operationId": "getPurse",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "user_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "purse by user",
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "balance": {
                                                    "type": "number"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "number"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "email": {
                                                            "type": "string"
                                                        },
                                                        "document": {
                                                            "type": "string"
                                                        },
                                                        "profile": {
                                                            "type": "string"
                                                        },
                                                        "status": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}